requirements.txt
numpy
tifffile
scipy
scikit-image
Pillow
